+++ sample-pscountdown.ps1
@@ -1,6 +1,6 @@
 <#
-A proof-of-concept script that adds an action at the end of the PSCountdownTimer
-This requires a Windows platform.
+A proof-of-concept script that adds an action at the end of the PSCountdownTimer.
+This sample requires a Windows platform.
 #>
 Clear-Host
 Start-PSCountdownTimer -seconds 60 -message "The PowerShell magic begins in " -FontSize 64 -Color SpringGreen
@@ -10,8 +10,9 @@ Do {
 
 Write-Host "Are you ready for some PowerShell?" -ForegroundColor magenta -BackgroundColor gray
 
-Add-Type –AssemblyName PresentationCore
+Add-Type -AssemblyName PresentationCore
 $filename = "c:\work\01-Start.mp3"
+
 #the media player launches with no UI. Use the object's methods to control it.
 $global:MediaPlayer = New-Object System.Windows.Media.MediaPlayer
 $global:MediaPlayer.Open($filename)